requirements.txt
ultralytics
opencv-python
torch
torchvision
scikit-learn
Pillow
deep_sort_realtime